refactor(breakpoints store): abstract breakpoint creation to separate method
