+++ src/BreakpointsStore/utils/breakpointFactory/index.ts
@@ -1,0 +1,1 @@
+export { breakpointFactory } from './breakpointFactory.util'
